backend/requirements.txt
opencv-python 
numpy 
scikit-learn
requests
python-dotenv
web3 
py-solc-x